[UPDATE] move category to exam package
