+++ src/main/java/com/depa/exam/dto/QuestionDTO.java
@@ -1,8 +1,6 @@
 package com.depa.exam.dto;
 
-import com.depa.category.dto.CategoryDTO;
-import com.depa.category.dto.impl.CategoryDTOImpl;
-import com.depa.category.model.Category;
+import com.depa.exam.dto.impl.CategoryDTOImpl;
 import com.depa.exam.model.question.Attribute;
 import com.depa.exam.model.question.Choice;
 import com.depa.exam.model.question.Question;
