+++ src/main/java/com/depa/exam/model/question/SubjectiveQuestion.java
@@ -1,7 +1,6 @@
 package com.depa.exam.model.question;
 
-import com.depa.category.dto.CategoryDTO;
-import com.depa.category.dto.impl.CategoryDTOImpl;
+import com.depa.exam.dto.impl.CategoryDTOImpl;
 
 import java.util.List;
