+++ src/test/java/com/depa/exam/service/QuestionServiceTest.java
@@ -1,8 +1,7 @@
 package com.depa.exam.service;
 
-import com.depa.category.dto.CategoryDTO;
-import com.depa.category.dto.impl.CategoryDTOImpl;
-import com.depa.category.model.Category;
+import com.depa.exam.dto.CategoryDTO;
+import com.depa.exam.dto.impl.CategoryDTOImpl;
 import com.depa.exam.dto.QuestionDTO;
 import com.depa.exam.model.question.Choice;
 import com.depa.exam.model.question.ObjectiveQuestion;
